refactor debug
